pull out PropagateNullable tests
