+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/optimizer/OptimizerRulesTests.java
@@ -86,105 +86,6 @@ public class OptimizerRulesTests extends ESTestCase {
     private static final Expression DUMMY_EXPRESSION =
         new org.elasticsearch.xpack.esql.core.optimizer.OptimizerRulesTests.DummyBooleanExpression(EMPTY, 0);
     //
-    // Null folding
-
-    //
-    // Propagate nullability (IS NULL / IS NOT NULL)
-    //
-
-    // a IS NULL AND a IS NOT NULL => false
-    public void testIsNullAndNotNull() {
-        FieldAttribute fa = getFieldAttribute();
-
-        And and = new And(EMPTY, new IsNull(EMPTY, fa), new IsNotNull(EMPTY, fa));
-        assertEquals(FALSE, new OptimizerRules.PropagateNullable().rule(and));
-    }
-
-    // a IS NULL AND b IS NOT NULL AND c IS NULL AND d IS NOT NULL AND e IS NULL AND a IS NOT NULL => false
-    public void testIsNullAndNotNullMultiField() {
-        FieldAttribute fa = getFieldAttribute();
-
-        And andOne = new And(EMPTY, new IsNull(EMPTY, fa), new IsNotNull(EMPTY, getFieldAttribute()));
-        And andTwo = new And(EMPTY, new IsNull(EMPTY, getFieldAttribute()), new IsNotNull(EMPTY, getFieldAttribute()));
-        And andThree = new And(EMPTY, new IsNull(EMPTY, getFieldAttribute()), new IsNotNull(EMPTY, fa));
-
-        And and = new And(EMPTY, andOne, new And(EMPTY, andThree, andTwo));
-
-        assertEquals(FALSE, new OptimizerRules.PropagateNullable().rule(and));
-    }
-
-    // a IS NULL AND a > 1 => a IS NULL AND false
-    public void testIsNullAndComparison() {
-        FieldAttribute fa = getFieldAttribute();
-        IsNull isNull = new IsNull(EMPTY, fa);
-
-        And and = new And(EMPTY, isNull, greaterThanOf(fa, ONE));
-        assertEquals(new And(EMPTY, isNull, nullOf(BOOLEAN)), new PropagateNullable().rule(and));
-    }
-
-    // a IS NULL AND b < 1 AND c < 1 AND a < 1 => a IS NULL AND b < 1 AND c < 1 => a IS NULL AND b < 1 AND c < 1
-    public void testIsNullAndMultipleComparison() {
-        FieldAttribute fa = getFieldAttribute();
-        IsNull isNull = new IsNull(EMPTY, fa);
-
-        And nestedAnd = new And(EMPTY, lessThanOf(getFieldAttribute("b"), ONE), lessThanOf(getFieldAttribute("c"), ONE));
-        And and = new And(EMPTY, isNull, nestedAnd);
-        And top = new And(EMPTY, and, lessThanOf(fa, ONE));
-
-        Expression optimized = new PropagateNullable().rule(top);
-        Expression expected = new And(EMPTY, and, nullOf(BOOLEAN));
-        assertEquals(Predicates.splitAnd(expected), Predicates.splitAnd(optimized));
-    }
-
-    // ((a+1)/2) > 1 AND a + 2 AND a IS NULL AND b < 3 => NULL AND NULL AND a IS NULL AND b < 3
-    public void testIsNullAndDeeplyNestedExpression() {
-        FieldAttribute fa = getFieldAttribute();
-        IsNull isNull = new IsNull(EMPTY, fa);
-
-        Expression nullified = new And(
-            EMPTY,
-            greaterThanOf(new Div(EMPTY, new Add(EMPTY, fa, ONE), TWO), ONE),
-            greaterThanOf(new Add(EMPTY, fa, TWO), ONE)
-        );
-        Expression kept = new And(EMPTY, isNull, lessThanOf(getFieldAttribute("b"), THREE));
-        And and = new And(EMPTY, nullified, kept);
-
-        Expression optimized = new PropagateNullable().rule(and);
-        Expression expected = new And(EMPTY, new And(EMPTY, nullOf(BOOLEAN), nullOf(BOOLEAN)), kept);
-
-        assertEquals(Predicates.splitAnd(expected), Predicates.splitAnd(optimized));
-    }
-
-    // a IS NULL OR a IS NOT NULL => no change
-    // a IS NULL OR a > 1 => no change
-    public void testIsNullInDisjunction() {
-        FieldAttribute fa = getFieldAttribute();
-
-        Or or = new Or(EMPTY, new IsNull(EMPTY, fa), new IsNotNull(EMPTY, fa));
-        Filter dummy = new Filter(EMPTY, relation(), or);
-        LogicalPlan transformed = new PropagateNullable().apply(dummy);
-        assertSame(dummy, transformed);
-        assertEquals(or, ((Filter) transformed).condition());
-
-        or = new Or(EMPTY, new IsNull(EMPTY, fa), greaterThanOf(fa, ONE));
-        dummy = new Filter(EMPTY, relation(), or);
-        transformed = new PropagateNullable().apply(dummy);
-        assertSame(dummy, transformed);
-        assertEquals(or, ((Filter) transformed).condition());
-    }
-
-    // a + 1 AND (a IS NULL OR a > 3) => no change
-    public void testIsNullDisjunction() {
-        FieldAttribute fa = getFieldAttribute();
-        IsNull isNull = new IsNull(EMPTY, fa);
-
-        Or or = new Or(EMPTY, isNull, greaterThanOf(fa, THREE));
-        And and = new And(EMPTY, new Add(EMPTY, fa, ONE), or);
-
-        assertEquals(and, new PropagateNullable().rule(and));
-    }
-
-    //
     // Like / Regex
     //
     public void testMatchAllLikeToExist() {
@@ -270,9 +171,6 @@ public class OptimizerRulesTests extends ESTestCase {
         return new IsNull(EMPTY, field);
     }
 
-    private Literal nullOf(DataType dataType) {
-        return new Literal(Source.EMPTY, null, dataType);
-    }
     //
     // Logical simplifications
     //
